+++ src/commonMain/kotlin/com/insanusmokrassar/sdi/HelperTools.kt
@@ -2,16 +2,25 @@ package com.insanusmokrassar.sdi
 
 import kotlinx.serialization.ImplicitReflectionSerializer
 import kotlinx.serialization.json.Json
+import kotlinx.serialization.modules.SerializersModuleBuilder
 
 internal val nonStrictJson = Json {
     isLenient = true
     ignoreUnknownKeys = true
     serializeSpecialFloatingPointValues = true
-    useArrayPolymorphism = true
 }
 
 @ImplicitReflectionSerializer
-fun Json.loadModule(json: String): Module = parse(ModuleDeserializerStrategy, json)
+private val ModuleDeserializerStrategyWithNullOptional = ModuleDeserializerStrategy()
 
 @ImplicitReflectionSerializer
-fun loadModule(json: String): Module = nonStrictJson.loadModule(json)
+fun Json.loadModule(
+    json: String,
+    moduleBuilder: (SerializersModuleBuilder.() -> Unit)? = null
+): Module = parse(if (moduleBuilder != null) ModuleDeserializerStrategy(moduleBuilder) else ModuleDeserializerStrategyWithNullOptional, json)
+
+@ImplicitReflectionSerializer
+fun loadModule(
+    json: String,
+    moduleBuilder: (SerializersModuleBuilder.() -> Unit)? = null
+): Module = nonStrictJson.loadModule(json, moduleBuilder)