+++ src/commonMain/kotlin/com/insanusmokrassar/sdi/ModuleDeserializerStrategy.kt
@@ -5,17 +5,19 @@ import kotlinx.serialization.*
 import kotlinx.serialization.builtins.MapSerializer
 import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.json.*
+import kotlinx.serialization.modules.SerializersModuleBuilder
 
 @ImplicitReflectionSerializer
-internal object ModuleDeserializerStrategy : DeserializationStrategy<Module> {
+internal class ModuleDeserializerStrategy(
+    private val moduleBuilder: (SerializersModuleBuilder.() -> Unit)? = null
+) : DeserializationStrategy<Module> {
     private val internalSerializer = MapSerializer(String.serializer(), ContextSerializer(Any::class))
     override val descriptor: SerialDescriptor
         get() = internalSerializer.descriptor
 
-    @InternalSerializationApi
     override fun deserialize(decoder: Decoder): Module {
         val json = JsonObjectSerializer.deserialize(decoder)
-        val jsonSerialFormat = createModuleBasedOnConfigRoot(json)
+        val jsonSerialFormat = createModuleBasedOnConfigRoot(json, moduleBuilder, decoder.context)
         val resultJson = JsonObject(
             json.keys.associateWith { JsonPrimitive(it) }
         )