+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # SDI changelogs
 
+## 0.3.0
+
+* Versions:
+    * `Kotlin`: `1.3.70` -> `1.3.71`
+* Common:
+    * Now it is possible to pass own lambda with receiver `SerializersModuleBuilder` to customize deserialization
+    * Now it will correctly resolve objects which was not previously registered
+    * By default, for modules loading will be used context from `Json`, passed as receiver
+
 ## 0.2.0
 
 * `Kotlin`: `1.3.61` -> `1.3.70`